Reduce renders in password when adding a user
